add useVideoAPI hook and remove unused hooks
