--- src/providers/Video/index.js
@@ -1,1 +1,0 @@
-export { default } from './Video.provider';
