+++ src/utils/apis/index.js
@@ -1,0 +1,1 @@
+export { default as VideoAPI } from './videoApi';
