+++ src/utils/constants.js
@@ -1,3 +1,4 @@
 const AUTH_STORAGE_KEY = 'wa_cert_authenticated';
+const YOUTUBE_API_KEY = '';
 
-export { AUTH_STORAGE_KEY };
+export { AUTH_STORAGE_KEY, YOUTUBE_API_KEY };
